# batch-shuffle-interface-poc