--- a/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/writer/ResultPartitionWriter.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/api/writer/ResultPartitionWriter.java
@@ -19,23 +19,26 @@
 package org.apache.flink.runtime.io.network.api.writer;
 
 import org.apache.flink.runtime.checkpoint.channel.ChannelStateReader;
+import org.apache.flink.runtime.event.AbstractEvent;
 import org.apache.flink.runtime.io.AvailabilityProvider;
-import org.apache.flink.runtime.io.network.buffer.BufferBuilder;
-import org.apache.flink.runtime.io.network.buffer.BufferConsumer;
+import org.apache.flink.runtime.io.network.partition.BufferAvailabilityListener;
 import org.apache.flink.runtime.io.network.partition.ResultPartitionID;
 import org.apache.flink.runtime.io.network.partition.ResultSubpartition;
+import org.apache.flink.runtime.io.network.partition.ResultSubpartitionView;
+import org.apache.flink.runtime.metrics.groups.TaskIOMetricGroup;
 
 import javax.annotation.Nullable;
 
 import java.io.IOException;
+import java.nio.ByteBuffer;
 
 /**
- * A buffer-oriented runtime result writer API for producing results.
+ * A record-oriented runtime result writer API for producing results.
  *
- * <p>If {@link ResultPartitionWriter#close()} is called before {@link ResultPartitionWriter#fail(Throwable)} or
- * {@link ResultPartitionWriter#finish()}, it abruptly triggers failure and cancellation of production.
- * In this case {@link ResultPartitionWriter#fail(Throwable)} still needs to be called afterwards to fully release
- * all resources associated the the partition and propagate failure cause to the consumer if possible.
+ * <p>If {@link ResultPartitionWriter#close()} is called before {@link ResultPartitionWriter#release(Throwable)}}
+ * or {@link ResultPartitionWriter#finish()}, it abruptly triggers failure and cancellation of production. In this
+ * case {@link ResultPartitionWriter#release(Throwable)}} still needs to be called afterwards to fully release all
+ * resources associated the the partition and propagate failure cause to the consumer if possible.
  */
 public interface ResultPartitionWriter extends AutoCloseable, AvailabilityProvider {
 
@@ -57,45 +60,23 @@ public interface ResultPartitionWriter extends AutoCloseable, AvailabilityProvid
 	int getNumTargetKeyGroups();
 
 	/**
-	 * Requests a {@link BufferBuilder} from this partition for writing data.
+	 * This is used to send regular records to the given channel.
 	 */
-	BufferBuilder getBufferBuilder(int targetChannel) throws IOException, InterruptedException;
-
-
-	/**
-	 * Try to request a {@link BufferBuilder} from this partition for writing data.
-	 *
-	 * <p>Returns <code>null</code> if no buffer is available or the buffer provider has been destroyed.
-	 */
-	BufferBuilder tryGetBufferBuilder(int targetChannel) throws IOException;
+	void writerRecord(ByteBuffer record, int targetChannel, boolean isBroadcastSelector) throws IOException, InterruptedException;
 
 	/**
-	 * Adds the bufferConsumer to the subpartition with the given index.
-	 *
-	 * <p>This method takes the ownership of the passed {@code bufferConsumer} and thus is responsible for releasing
-	 * it's resources.
-	 *
-	 * <p>To avoid problems with data re-ordering, before adding new {@link BufferConsumer} the previously added one
-	 * the given {@code subpartitionIndex} must be marked as {@link BufferConsumer#isFinished()}.
-	 *
-	 * @return true if operation succeeded and bufferConsumer was enqueued for consumption.
+	 * This is used to broadcast regular records or streaming Watermarks in-band with records to all channels.
 	 */
-	boolean addBufferConsumer(BufferConsumer bufferConsumer, int subpartitionIndex, boolean isPriorityEvent) throws IOException;
+	void broadcastWrite(ByteBuffer record, boolean isBroadcastSelector) throws IOException, InterruptedException;
 
 	/**
-	 * Adds the bufferConsumer to the subpartition with the given index.
-	 *
-	 * <p>This method takes the ownership of the passed {@code bufferConsumer} and thus is responsible for releasing
-	 * it's resources.
-	 *
-	 * <p>To avoid problems with data re-ordering, before adding new {@link BufferConsumer} the previously added one
-	 * the given {@code subpartitionIndex} must be marked as {@link BufferConsumer#isFinished()}.
-	 *
-	 * @return true if operation succeeded and bufferConsumer was enqueued for consumption.
+	 * This is used to broadcast event to all channels.
 	 */
-	default boolean addBufferConsumer(BufferConsumer bufferConsumer, int subpartitionIndex) throws IOException {
-		return addBufferConsumer(bufferConsumer, subpartitionIndex, false);
-	}
+	void broadcastEvent(AbstractEvent event, boolean isPriorityEvent) throws IOException;
+
+	void setMetricGroup(TaskIOMetricGroup metrics);
+
+	void setConsumableNotifier(Runnable consumableNotifier);
 
 	/**
 	 * Returns the subpartition with the given index.
@@ -103,25 +84,19 @@ public interface ResultPartitionWriter extends AutoCloseable, AvailabilityProvid
 	ResultSubpartition getSubpartition(int subpartitionIndex);
 
 	/**
-	 * Manually trigger consumption from enqueued {@link BufferConsumer BufferConsumers} in all subpartitions.
+	 * Manually trigger consumption of all subpartitions.
 	 */
 	void flushAll();
 
 	/**
-	 * Manually trigger consumption from enqueued {@link BufferConsumer BufferConsumers} in one specified subpartition.
+	 * Manually trigger consumption of the specified subpartition.
 	 */
 	void flush(int subpartitionIndex);
 
 	/**
-	 * Fail the production of the partition.
-	 *
-	 * <p>This method propagates non-{@code null} failure causes to consumers on a best-effort basis. This call also
-	 * leads to the release of all resources associated with the partition. Closing of the partition is still needed
-	 * afterwards if it has not been done before.
-	 *
-	 * @param throwable failure cause
+	 * Returns the read view of the requested subpartition.
 	 */
-	void fail(@Nullable Throwable throwable);
+	ResultSubpartitionView createSubpartitionView(int subpartitionIndex, BufferAvailabilityListener listener) throws IOException;
 
 	/**
 	 * Successfully finish the production of the partition.
@@ -129,4 +104,13 @@ public interface ResultPartitionWriter extends AutoCloseable, AvailabilityProvid
 	 * <p>Closing of partition is still needed afterwards.
 	 */
 	void finish() throws IOException;
+
+	/**
+	 * Releases the result partition writer.
+	 */
+	void release(@Nullable Throwable cause);
+
+	boolean isFinished();
+
+	boolean isReleased();
 }