--- a/flink-runtime/src/main/java/org/apache/flink/runtime/operators/DataSourceTask.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/operators/DataSourceTask.java
@@ -233,7 +233,6 @@ public class DataSourceTask<OT> extends AbstractInvokable {
 				BatchTask.logAndThrowException(ex, this);
 			}
 		} finally {
-			BatchTask.clearWriters(eventualOutputs);
 			// --------------------------------------------------------------------
 			// Closing
 			// --------------------------------------------------------------------