--- a/flink-tests/src/test/java/org/apache/flink/test/runtime/ShuffleCompressionITCase.java
+++ b/flink-tests/src/test/java/org/apache/flink/test/runtime/ShuffleCompressionITCase.java
@@ -167,7 +167,6 @@ public class ShuffleCompressionITCase {
 				writer.broadcastEmit(RECORD_TO_SEND);
 			}
 			writer.flushAll();
-			writer.clearBuffers();
 		}
 	}